+++ LEDMatrix/Table_Mark_Estes34/PatternWave.h
@@ -39,8 +39,8 @@ private:
 
     uint8_t scale = 256 / MATRIX_WIDTH;
 
-    uint8_t maxX = MATRIX_WIDTH - 1;
-    uint8_t maxY = MATRIX_HEIGHT - 1;
+    uint16_t maxX = MATRIX_WIDTH - 1;
+    uint16_t maxY = MATRIX_HEIGHT - 1;
 
     uint8_t waveCount = 1;
